Move README.en.md to README.md due to Github bug
README.en.md is not displayed on the landing page of the Github
repository. This commit address this issue temporarily until a more
persistent solution can be devised. That, or we decide to relegate
the Github repository to mirror status.

Gitea, an open source competitor, gets this UX bug right.